feat: 코드 리뷰 반영
- 주석 삭제
- TokenData 삭제
- Repository에서 try-except 대신 len을 이용하여 return None 처리
- CatRepository에서 UserRepository를 사용하여 중복 코드 제거
