+++ api/deps.py
@@ -3,18 +3,8 @@ from core.config import settings
 from fastapi import Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordBearer
 from database import SessionLocal
-from schema.login import TokenData
 
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/login/")
-
-
-# def get_current_user_id(token: str = Depends(oauth2_scheme)):
-#     decoded_token = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
-#     if not decoded_token["sub"]:
-#         raise HTTPException(status_code=404, detail="user의 id값을 찾을 수 없습니다.")
-#     id = int(decoded_token["sub"])
-
-#     return id
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/login")
 
 
 def get_current_user_id(token: str = Depends(oauth2_scheme)) -> str:
@@ -28,10 +18,9 @@ def get_current_user_id(token: str = Depends(oauth2_scheme)) -> str:
         user_id: str = payload.get("sub")
         if user_id is None:
             raise credentials_exception
-        token_data = TokenData(user_id=user_id)
     except JWTError:
         raise credentials_exception
-    return token_data.user_id
+    return user_id
 
 
 def get_db():
